refactor: split chat box into different panels
